Fixed the target validation check
  Removed deprecated LUTs
