+++ moreHUD.vcxproj.filters
@@ -260,8 +260,6 @@
     <ClCompile Include="AHZScaleformHook.cpp" />
     <ClCompile Include="AHZUtility.cpp" />
     <ClCompile Include="AHZWeaponInfo.cpp" />
-    <ClCompile Include="FoodLUT.cpp" />
-    <ClCompile Include="IngredientLUT.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="..\skse64_common\SafeWrite.cpp" />
     <ClCompile Include="..\skse64_common\Utilities.cpp">
@@ -450,12 +448,11 @@
     <ClCompile Include="..\skse64\Serialization.cpp">
       <Filter>skse64\SKSE64 Dependencies</Filter>
     </ClCompile>
-    <ClCompile Include="ShrineLUT.cpp" />
     <ClCompile Include="..\skse64\InternalTasks.cpp">
       <Filter>skse64\SKSE64 Dependencies</Filter>
     </ClCompile>
     <ClCompile Include="AHZConsole.cpp" />
-    <ClCompile Include="AHZVM.cpp" />
+    <ClCompile Include="AHZACTIHelper.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\skse64\GameAPI.h">
@@ -472,12 +469,9 @@
     <ClInclude Include="AHZScaleformHook.h" />
     <ClInclude Include="AHZUtility.h" />
     <ClInclude Include="AHZWeaponInfo.h" />
-    <ClInclude Include="FoodLUT.h" />
-    <ClInclude Include="IngredientLUT.h" />
     <ClInclude Include="resource2.h" />
-    <ClInclude Include="ShrineLUT.h" />
     <ClInclude Include="AHZConsole.h" />
-    <ClInclude Include="AHZVM.h" />
+    <ClInclude Include="AHZACTIHelper.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="exports.def" />
